requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
resend==0.7.0
python-dotenv==1.0.0
